Bug fixes, upgrades in it's internal design
The platform had some details that needed to be fixed, only in the frontend side of things, stuff having to do with css files and such, to be more robust and also ridden of errors that may end up ruining the user experience, very few things get an user to think a page is terrible than visual bugs.

The platform as it stands is just a prototype for a larger concept that may or may not grow, this page is just a demo, must reiterate that.